Rename package from wavefront_plugin to wavefront
This ensures there is consistency between the package folder structure
and the package name. This is important for users who are using the SDK
directly rather than just the CLI based tool

There is no functionality changes in this PR - it is merely mechanical
renaming
