+++ main.go
@@ -9,7 +9,7 @@ import (
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() terraform.ResourceProvider {
-			return wavefront_plugin.Provider()
+			return wavefront.Provider()
 		},
 	})
 }
