+++ wavefront/resource_cloud_integration_app_dynamics_test.go
@@ -1,4 +1,4 @@
-package wavefront_plugin
+package wavefront
 
 import (
 	"fmt"
