+++ wavefront/resource_cloud_integration_aws_external_id_test.go
@@ -1,4 +1,4 @@
-package wavefront_plugin
+package wavefront
 
 import (
 	"fmt"
