+++ wavefront/resource_cloud_integration_gcp.go
@@ -1,4 +1,4 @@
-package wavefront_plugin
+package wavefront
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
